patch by Gaby: changed --athlower to accept floats

+++ include/lame.h
@@ -129,7 +129,7 @@ typedef struct  {
   int ATHshort;                   /* only use ATH for short blocks */
   int noATH;                      /* disable ATH */
   int ATHtype;                    /* select ATH formula */
-  int ATHlower;                   /* lower ATH by this many db */
+  float ATHlower;                 /* lower ATH by this many db */
   int cwlimit;                    /* predictability limit */
   int allow_diff_short;           /* allow blocktypes to differ between channels ? */
   int useTemporal;                /* use temporal masking effect */
